Introduce CJS exports
ES modules cannot be imported (required) to a CJS modules. It's a pity
to be unable to use a module like this in a CJS project, especially
because offering a CJS build output next to the ES one is very easy.

The final solution would probably somehow make use of `@ehmicky/dev-tasks`,
but it'd mean more work. Consumers of this package won't notice any
difference, if it's done later.